Added session cookie on request to TFL
There appears to be a problem with one of the TFL servers causing to sent incorrect data. It also appears that users are sticky cookied to one machine, so I have added this cookie to the request so it always hits the same server. Not sure if the cookie sends it to the correct server or not, needs testing against real life data to check.